fix: アップロード処理を Component に分割
